Fix GitHub query for keyword in topic, name or description
Search for a keyword in either topic, name, or description is not possible at once, but has to be done separately, i.e. once for topic, and then for name or description.
As this doubles the number of queries and Taskwarrior has a lot of tools associated, we have to take heed of rate limiting.
GitHub allows 30 requests per minute for the search API which delivers results in chunks of 30 items.
This resolves approximately to 1 item every 0.06 seconds.
Added a pause which should ensure we stay below this limit.